Fixed introduced vardecimal parsing errors. Tests green again.

+++ src/OrcaMDF.Core/Engine/SqlTypes/SqlInt.cs
@@ -50,7 +50,7 @@ namespace OrcaMDF.Core.Engine.SqlTypes
 				if (value.Length != 4)
 					throw new ArgumentException("Invalid value length: " + value.Length);
 
-				return BitConverter.ToInt16(value, 0);
+				return BitConverter.ToInt32(value, 0);
 			}
 		}
 	}
